Refactoring example handling to compile all source YAML or JSON
