+++ .gitignore
@@ -14,3 +14,5 @@
 # Dependency directories (remove the comment below to include it)
 # vendor/
 
+# Output from example testing
+*.out.json
