+++ examples_test.go
@@ -3,46 +3,94 @@ package gobl_test
 import (
 	"encoding/json"
 	"fmt"
+	"io/ioutil"
+	"os"
+	"path/filepath"
+	"strings"
+	"testing"
 
 	"github.com/invopop/gobl"
-	"github.com/invopop/gobl/note"
-	"github.com/invopop/gobl/uuid"
+	"github.com/invopop/yaml"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 )
 
-func ExampleNewEnvelope_complete() {
-	// Prepare a new Envelope with a region
-	env := gobl.NewEnvelope()
-	env.Head.UUID = uuid.MustParse("871c1e6a-8b5c-11ec-af5f-3e7e00ce5635")
+var skipExamplePaths = []string{
+	"build/",
+	".out.",
+	".github",
+}
 
-	// Prepare a payload and insert
-	msg := &note.Message{
-		Content: "sample message content",
-	}
-	if err := env.Insert(msg); err != nil {
-		panic(err.Error())
-	}
-	if err := env.Validate(); err != nil {
-		panic(err.Error())
-	}
+// TestConvertExamplesToJSON finds all of the `.json` and `.yaml` files in the
+// package and attempts to convert the to JSON Envelopes.
+func TestConvertExamplesToJSON(t *testing.T) {
+	// Find all .yaml files in subdirectories
+	var files []string
+	err := filepath.Walk("./", func(path string, info os.FileInfo, err error) error {
+		switch filepath.Ext(path) {
+		case ".yaml", ".json":
+			for _, skip := range skipExamplePaths {
+				if strings.Contains(path, skip) {
+					return nil
+				}
+			}
+			files = append(files, path)
+		}
+		return nil
+	})
+	require.NoError(t, err)
 
-	data, err := json.MarshalIndent(env, "", "\t")
+	for _, path := range files {
+		assert.NoError(t, processFile(t, path))
+	}
+}
+
+func processFile(t *testing.T, path string) error {
+	t.Helper()
+	t.Logf("processing file: %v", path)
+
+	// attempt to load and convert
+	data, err := ioutil.ReadFile(path)
 	if err != nil {
-		panic(err.Error())
+		return fmt.Errorf("reading file: %w", err)
 	}
-	fmt.Printf("%v\n", string(data))
-	// Output:
-	// {
-	// 	"$schema": "https://gobl.org/draft-0/envelope",
-	// 	"head": {
-	// 		"uuid": "871c1e6a-8b5c-11ec-af5f-3e7e00ce5635",
-	// 		"dig": {
-	// 			"alg": "sha256",
-	// 			"val": "7d539c46ca03a4ecb1fcc4cb00d2ada34275708ee326caafee04d9dcfed862ee"
-	// 		}
-	// 	},
-	// 	"doc": {
-	// 		"$schema": "https://gobl.org/draft-0/note/message",
-	// 		"content": "sample message content"
-	// 	}
-	// }
+
+	var env *gobl.Envelope
+	if strings.Contains(path, ".env.") {
+		// Handle Envelopes
+		env = new(gobl.Envelope)
+		if err := yaml.Unmarshal(data, env); err != nil {
+			return fmt.Errorf("invalid contents: %w", err)
+		}
+		if err := env.Calculate(); err != nil {
+			return fmt.Errorf("failed to complete: %w", err)
+		}
+	} else {
+		// Handle documents
+		doc := new(gobl.Document)
+		if err := yaml.Unmarshal(data, doc); err != nil {
+			return fmt.Errorf("invalid contents: %w", err)
+		}
+		env, err = gobl.Envelop(doc)
+		if err != nil {
+			return fmt.Errorf("failed to envelop: %w", err)
+		}
+	}
+
+	if err := env.Sign(testKey); err != nil {
+		return fmt.Errorf("failed to sign the doc: %w", err)
+	}
+
+	// Output to the filesystem (.out.json is defined in .gitignore)
+	np := strings.TrimSuffix(path, filepath.Ext(path)) + ".out.json"
+	out, err := json.MarshalIndent(env, "", "	")
+	if err != nil {
+		return fmt.Errorf("marshalling output: %w", err)
+	}
+	if err := ioutil.WriteFile(np, out, 0644); err != nil {
+		return fmt.Errorf("saving file data: %w", err)
+	}
+
+	t.Logf("wrote file: %v", np)
+	return nil
 }
