+++ note/examples/message.yaml
