+++ regimes/es/examples/invoice-es-nl-digital-b2c.yaml
