--- samples/.gitignore
@@ -1,2 +1,0 @@
-*.json
-!output/**
