--- samples/output/es/invoice-es-es-freelance.json
@@ -1,126 +1,0 @@
-{
-	"$schema": "https://gobl.org/draft-0/envelope",
-	"head": {
-		"uuid": "25200bc1-ebc7-11ec-9155-e2c0624703d9",
-		"dig": {
-			"alg": "sha256",
-			"val": "1779c452e2aa474b40577ca817d7c72f13f33da06a51af7c267421084fa1b86b"
-		}
-	},
-	"doc": {
-		"$schema": "https://gobl.org/draft-0/bill/invoice",
-		"code": "SAMPLE-001",
-		"currency": "EUR",
-		"issue_date": "2022-02-01",
-		"supplier": {
-			"tax_id": {
-				"country": "ES",
-				"code": "58384285G"
-			},
-			"name": "MªF. Services",
-			"people": [
-				{
-					"name": {
-						"given": "MARIA FRANCISCA",
-						"surname": "MONTERO",
-						"surname2": "ESTEBAN"
-					}
-				}
-			],
-			"addresses": [
-				{
-					"num": "9",
-					"street": "CAMÍ MADRID",
-					"locality": "CANENA",
-					"region": "JAÉN",
-					"code": "23480",
-					"country": "ES"
-				}
-			],
-			"emails": [
-				{
-					"addr": "billing@example.com"
-				}
-			]
-		},
-		"customer": {
-			"tax_id": {
-				"country": "ES",
-				"code": "54387763P"
-			},
-			"name": "Sample Consumer"
-		},
-		"lines": [
-			{
-				"i": 1,
-				"quantity": "20",
-				"item": {
-					"name": "Development services",
-					"price": "90.00",
-					"unit": "h"
-				},
-				"sum": "1800.00",
-				"discounts": [
-					{
-						"percent": "10%",
-						"amount": "180.00",
-						"reason": "Special discount"
-					}
-				],
-				"taxes": [
-					{
-						"cat": "VAT",
-						"rate": "standard",
-						"percent": "21.0%"
-					},
-					{
-						"cat": "IRPF",
-						"percent": "15.0%"
-					}
-				],
-				"total": "1620.00"
-			}
-		],
-		"totals": {
-			"sum": "1620.00",
-			"total": "1620.00",
-			"taxes": {
-				"categories": [
-					{
-						"code": "VAT",
-						"rates": [
-							{
-								"key": "standard",
-								"base": "1620.00",
-								"percent": "21.0%",
-								"amount": "340.20"
-							}
-						],
-						"base": "1620.00",
-						"amount": "340.20"
-					},
-					{
-						"code": "IRPF",
-						"retained": true,
-						"rates": [
-							{
-								"base": "1620.00",
-								"percent": "15.0%",
-								"amount": "243.00"
-							}
-						],
-						"base": "1620.00",
-						"amount": "243.00"
-					}
-				],
-				"sum": "97.20"
-			},
-			"tax": "97.20",
-			"total_with_tax": "1717.20",
-			"payable": "1717.20"
-		}
-	},
-	"sigs": [
-		"eyJhbGciOiJFUzI1NiIsImtpZCI6IjgzNDdjNGQwLTdjZDQtNGMyMi05OWY2LTg5MWI1NDRlMWMxZCJ9.eyJ1dWlkIjoiMjUyMDBiYzEtZWJjNy0xMWVjLTkxNTUtZTJjMDYyNDcwM2Q5IiwiZGlnIjp7ImFsZyI6InNoYTI1NiIsInZhbCI6IjE3NzljNDUyZTJhYTQ3NGI0MDU3N2NhODE3ZDdjNzJmMTNmMzNkYTA2YTUxYWY3YzI2NzQyMTA4NGZhMWI4NmIifX0.bjyX78M_jzsFQQHO5D5nLutD0GlS-dDc8ao-i9XiMZDR9xB0egAzrNPNtY4sMFQYQyja_dauNlRFz35Nqwtk8A"
-	]
-}
