--- samples/output/es/invoice-es-es-vateqs-provider.json
@@ -1,134 +1,0 @@
-{
-	"$schema": "https://gobl.org/draft-0/envelope",
-	"head": {
-		"uuid": "25208e57-ebc7-11ec-9155-e2c0624703d9",
-		"dig": {
-			"alg": "sha256",
-			"val": "9f77160ae8757ede845eaa3bcf7a4fa9aafbb05c2bfac22929c08a04399ce274"
-		}
-	},
-	"doc": {
-		"$schema": "https://gobl.org/draft-0/bill/invoice",
-		"code": "SAMPLE-001",
-		"currency": "EUR",
-		"tax": {
-			"prices_include": "VAT"
-		},
-		"issue_date": "2022-02-01",
-		"supplier": {
-			"tax_id": {
-				"country": "ES",
-				"code": "B98602642"
-			},
-			"name": "Provide One S.L.",
-			"addresses": [
-				{
-					"num": "42",
-					"street": "Calle Pradillo",
-					"locality": "Madrid",
-					"region": "Madrid",
-					"code": "28002",
-					"country": "ES"
-				}
-			],
-			"emails": [
-				{
-					"addr": "billing@example.com"
-				}
-			]
-		},
-		"customer": {
-			"tax_id": {
-				"country": "ES",
-				"code": "54387763P"
-			},
-			"name": "Simple Goods Store",
-			"addresses": [
-				{
-					"num": "43",
-					"street": "Calle Mayor",
-					"locality": "",
-					"region": "Madrid",
-					"code": "28003"
-				}
-			]
-		},
-		"lines": [
-			{
-				"i": 1,
-				"quantity": "10",
-				"item": {
-					"name": "Mugs from provider",
-					"price": "10.00"
-				},
-				"sum": "100.00",
-				"taxes": [
-					{
-						"cat": "VAT",
-						"rate": "standard+eqs",
-						"percent": "21.0%",
-						"surcharge": "5.2%"
-					}
-				],
-				"total": "100.00"
-			},
-			{
-				"i": 2,
-				"quantity": "1",
-				"item": {
-					"name": "Delivery Costs",
-					"price": "10.00"
-				},
-				"sum": "10.00",
-				"taxes": [
-					{
-						"cat": "VAT",
-						"rate": "standard",
-						"percent": "21.0%"
-					}
-				],
-				"total": "10.00"
-			}
-		],
-		"totals": {
-			"sum": "110.00",
-			"tax_included": "19.10",
-			"total": "90.90",
-			"taxes": {
-				"categories": [
-					{
-						"code": "VAT",
-						"rates": [
-							{
-								"key": "standard+eqs",
-								"base": "82.64",
-								"percent": "21.0%",
-								"amount": "17.36",
-								"surcharge": {
-									"percent": "5.2%",
-									"amount": "4.29"
-								}
-							},
-							{
-								"key": "standard",
-								"base": "8.26",
-								"percent": "21.0%",
-								"amount": "1.74"
-							}
-						],
-						"base": "90.90",
-						"amount": "19.10",
-						"surcharge": "4.29"
-					}
-				],
-				"sum": "23.39"
-			},
-			"tax": "23.39",
-			"total_with_tax": "114.29",
-			"payable": "114.29"
-		}
-	},
-	"sigs": [
-		"eyJhbGciOiJFUzI1NiIsImtpZCI6IjgzNDdjNGQwLTdjZDQtNGMyMi05OWY2LTg5MWI1NDRlMWMxZCJ9.eyJ1dWlkIjoiMjUyMDhlNTctZWJjNy0xMWVjLTkxNTUtZTJjMDYyNDcwM2Q5IiwiZGlnIjp7ImFsZyI6InNoYTI1NiIsInZhbCI6IjlmNzcxNjBhZTg3NTdlZGU4NDVlYWEzYmNmN2E0ZmE5YWFmYmIwNWMyYmZhYzIyOTI5YzA4YTA0Mzk5Y2UyNzQifX0.0iMR-CQw5nmR5CcYmfGRQnshLN3HfzCdypgG_YH9SISAOEbeNp9KxanTKbmaKx7sM4006eXtI4QlBkJGhSTbAw"
-	]
-}
