--- samples/output/es/invoice-es-es-vateqs-retailer.json
@@ -1,86 +1,0 @@
-{
-	"$schema": "https://gobl.org/draft-0/envelope",
-	"head": {
-		"uuid": "2520c550-ebc7-11ec-9155-e2c0624703d9",
-		"dig": {
-			"alg": "sha256",
-			"val": "db527b9ef61fe2216c6bc898c3a4662263fb53be81a1da05e3bf8a51f0b66a21"
-		}
-	},
-	"doc": {
-		"$schema": "https://gobl.org/draft-0/bill/invoice",
-		"code": "SAMPLE-001",
-		"currency": "EUR",
-		"tax": {
-			"prices_include": "VAT"
-		},
-		"issue_date": "2022-02-01",
-		"supplier": {
-			"tax_id": {
-				"country": "ES",
-				"code": "54387763P"
-			},
-			"name": "Simple Goods Store",
-			"addresses": [
-				{
-					"num": "43",
-					"street": "Calle Mayor",
-					"locality": "",
-					"region": "Madrid",
-					"code": "28003"
-				}
-			]
-		},
-		"lines": [
-			{
-				"i": 1,
-				"quantity": "10",
-				"item": {
-					"name": "Mugs from provider",
-					"price": "16.00",
-					"meta": {
-						"source": "provider"
-					}
-				},
-				"sum": "160.00",
-				"taxes": [
-					{
-						"cat": "VAT",
-						"rate": "standard",
-						"percent": "21.0%"
-					}
-				],
-				"total": "160.00"
-			}
-		],
-		"totals": {
-			"sum": "160.00",
-			"tax_included": "27.77",
-			"total": "132.23",
-			"taxes": {
-				"categories": [
-					{
-						"code": "VAT",
-						"rates": [
-							{
-								"key": "standard",
-								"base": "132.23",
-								"percent": "21.0%",
-								"amount": "27.77"
-							}
-						],
-						"base": "132.23",
-						"amount": "27.77"
-					}
-				],
-				"sum": "27.77"
-			},
-			"tax": "27.77",
-			"total_with_tax": "160.00",
-			"payable": "160.00"
-		}
-	},
-	"sigs": [
-		"eyJhbGciOiJFUzI1NiIsImtpZCI6IjgzNDdjNGQwLTdjZDQtNGMyMi05OWY2LTg5MWI1NDRlMWMxZCJ9.eyJ1dWlkIjoiMjUyMGM1NTAtZWJjNy0xMWVjLTkxNTUtZTJjMDYyNDcwM2Q5IiwiZGlnIjp7ImFsZyI6InNoYTI1NiIsInZhbCI6ImRiNTI3YjllZjYxZmUyMjE2YzZiYzg5OGMzYTQ2NjIyNjNmYjUzYmU4MWExZGEwNWUzYmY4YTUxZjBiNjZhMjEifX0.hC64zQDo6kPw7IX5NMoDOrEFNQ_fg7uZ0ul4GxtuglL92ej_H8JBy7bP4IgZaTJQ6t7N4Ng3pVdd2N5iFUMVJg"
-	]
-}
