--- samples/output/es/invoice-es-es.env.json
@@ -1,92 +1,0 @@
-{
-	"$schema": "https://gobl.org/draft-0/envelope",
-	"head": {
-		"uuid": "2520e81b-ebc7-11ec-9155-e2c0624703d9",
-		"dig": {
-			"alg": "sha256",
-			"val": "8e9ddd2abcc870e0a080a511c8c4796a674633dbb534687f8d2f3f7351530348"
-		}
-	},
-	"doc": {
-		"$schema": "https://gobl.org/draft-0/bill/invoice",
-		"code": "SAMPLE-001",
-		"currency": "EUR",
-		"issue_date": "2022-02-01",
-		"supplier": {
-			"tax_id": {
-				"country": "ES",
-				"code": "B98602642"
-			},
-			"name": "Provide One S.L.",
-			"addresses": [
-				{
-					"num": "42",
-					"street": "Calle Pradillo",
-					"locality": "Madrid",
-					"region": "Madrid",
-					"code": "28002",
-					"country": "ES"
-				}
-			],
-			"emails": [
-				{
-					"addr": "billing@example.com"
-				}
-			]
-		},
-		"customer": {
-			"tax_id": {
-				"country": "ES",
-				"code": "54387763P"
-			},
-			"name": "Sample Consumer"
-		},
-		"lines": [
-			{
-				"i": 1,
-				"quantity": "10",
-				"item": {
-					"name": "Item being purchased",
-					"price": "100.00"
-				},
-				"sum": "1000.00",
-				"taxes": [
-					{
-						"cat": "VAT",
-						"rate": "standard",
-						"percent": "21.0%"
-					}
-				],
-				"total": "1000.00"
-			}
-		],
-		"totals": {
-			"sum": "1000.00",
-			"total": "1000.00",
-			"taxes": {
-				"categories": [
-					{
-						"code": "VAT",
-						"rates": [
-							{
-								"key": "standard",
-								"base": "1000.00",
-								"percent": "21.0%",
-								"amount": "210.00"
-							}
-						],
-						"base": "1000.00",
-						"amount": "210.00"
-					}
-				],
-				"sum": "210.00"
-			},
-			"tax": "210.00",
-			"total_with_tax": "1210.00",
-			"payable": "1210.00"
-		}
-	},
-	"sigs": [
-		"eyJhbGciOiJFUzI1NiIsImtpZCI6IjgzNDdjNGQwLTdjZDQtNGMyMi05OWY2LTg5MWI1NDRlMWMxZCJ9.eyJ1dWlkIjoiMjUyMGU4MWItZWJjNy0xMWVjLTkxNTUtZTJjMDYyNDcwM2Q5IiwiZGlnIjp7ImFsZyI6InNoYTI1NiIsInZhbCI6IjhlOWRkZDJhYmNjODcwZTBhMDgwYTUxMWM4YzQ3OTZhNjc0NjMzZGJiNTM0Njg3ZjhkMmYzZjczNTE1MzAzNDgifX0.8EU6EWJo57mW006uwLXEJYoiITB7ig86sifitF-4m_W61PYoG8sXd4gfNexHcIFUV4IrxhjxPjtEp-GQopuIbQ"
-	]
-}
