--- samples/output/es/invoice-es-es.json
@@ -1,123 +1,0 @@
-{
-	"$schema": "https://gobl.org/draft-0/envelope",
-	"head": {
-		"uuid": "2521180d-ebc7-11ec-9155-e2c0624703d9",
-		"dig": {
-			"alg": "sha256",
-			"val": "7b65a0c5db92771f65963ae7152d5e78c6115cb21c312cca9821716c2881c76f"
-		}
-	},
-	"doc": {
-		"$schema": "https://gobl.org/draft-0/bill/invoice",
-		"code": "SAMPLE-001",
-		"currency": "EUR",
-		"issue_date": "2022-02-01",
-		"supplier": {
-			"tax_id": {
-				"country": "ES",
-				"code": "B98602642"
-			},
-			"name": "Provide One S.L.",
-			"addresses": [
-				{
-					"num": "42",
-					"street": "Calle Pradillo",
-					"locality": "Madrid",
-					"region": "Madrid",
-					"code": "28002",
-					"country": "ES"
-				}
-			],
-			"emails": [
-				{
-					"addr": "billing@example.com"
-				}
-			]
-		},
-		"customer": {
-			"tax_id": {
-				"country": "ES",
-				"code": "54387763P"
-			},
-			"name": "Sample Consumer"
-		},
-		"lines": [
-			{
-				"i": 1,
-				"quantity": "20",
-				"item": {
-					"name": "Development services",
-					"price": "90.00",
-					"unit": "h"
-				},
-				"sum": "1800.00",
-				"discounts": [
-					{
-						"percent": "10%",
-						"amount": "180.00",
-						"reason": "Special discount"
-					}
-				],
-				"taxes": [
-					{
-						"cat": "VAT",
-						"rate": "standard",
-						"percent": "21.0%"
-					}
-				],
-				"total": "1620.00"
-			},
-			{
-				"i": 2,
-				"quantity": "1",
-				"item": {
-					"name": "Financial service",
-					"price": "10.00"
-				},
-				"sum": "10.00",
-				"taxes": [
-					{
-						"cat": "VAT",
-						"rate": "zero",
-						"percent": "0.0%"
-					}
-				],
-				"total": "10.00"
-			}
-		],
-		"totals": {
-			"sum": "1630.00",
-			"total": "1630.00",
-			"taxes": {
-				"categories": [
-					{
-						"code": "VAT",
-						"rates": [
-							{
-								"key": "standard",
-								"base": "1620.00",
-								"percent": "21.0%",
-								"amount": "340.20"
-							},
-							{
-								"key": "zero",
-								"base": "10.00",
-								"percent": "0.0%",
-								"amount": "0.00"
-							}
-						],
-						"base": "1630.00",
-						"amount": "340.20"
-					}
-				],
-				"sum": "340.20"
-			},
-			"tax": "340.20",
-			"total_with_tax": "1970.20",
-			"payable": "1970.20"
-		}
-	},
-	"sigs": [
-		"eyJhbGciOiJFUzI1NiIsImtpZCI6IjgzNDdjNGQwLTdjZDQtNGMyMi05OWY2LTg5MWI1NDRlMWMxZCJ9.eyJ1dWlkIjoiMjUyMTE4MGQtZWJjNy0xMWVjLTkxNTUtZTJjMDYyNDcwM2Q5IiwiZGlnIjp7ImFsZyI6InNoYTI1NiIsInZhbCI6IjdiNjVhMGM1ZGI5Mjc3MWY2NTk2M2FlNzE1MmQ1ZTc4YzYxMTVjYjIxYzMxMmNjYTk4MjE3MTZjMjg4MWM3NmYifX0.B8KwMAHqvHDsu4dOmvLbUdY6keA6Xq9kRObIZi4KGzgQPbjjR1nEROCYKKFH2nPwN4RqxorZNzsPi1dcbq8IAA"
-	]
-}
