--- samples/output/es/invoice-es-nl-b2b.json
@@ -1,127 +1,0 @@
-{
-	"$schema": "https://gobl.org/draft-0/envelope",
-	"head": {
-		"uuid": "25214235-ebc7-11ec-9155-e2c0624703d9",
-		"dig": {
-			"alg": "sha256",
-			"val": "2c16a009c575ceb862e94a6da9a8b90140b6aab6ea93b37d154fe196618e09cd"
-		}
-	},
-	"doc": {
-		"$schema": "https://gobl.org/draft-0/bill/invoice",
-		"code": "SAMPLE-X-002",
-		"currency": "EUR",
-		"tax": {
-			"prices_include": "VAT",
-			"schemes": [
-				"reverse-charge"
-			]
-		},
-		"issue_date": "2022-02-01",
-		"supplier": {
-			"tax_id": {
-				"country": "ES",
-				"code": "B98602642"
-			},
-			"name": "Provide One S.L.",
-			"addresses": [
-				{
-					"num": "42",
-					"street": "Calle Pradillo",
-					"locality": "Madrid",
-					"region": "Madrid",
-					"code": "28002",
-					"country": "ES"
-				}
-			],
-			"emails": [
-				{
-					"addr": "billing@example.com"
-				}
-			]
-		},
-		"customer": {
-			"tax_id": {
-				"country": "NL",
-				"code": "NL000099995B57"
-			},
-			"name": "Sample Consumer"
-		},
-		"lines": [
-			{
-				"i": 1,
-				"quantity": "10",
-				"item": {
-					"name": "Services exported",
-					"price": "20.00",
-					"unit": "day"
-				},
-				"sum": "200.00",
-				"taxes": [
-					{
-						"cat": "VAT",
-						"rate": "standard",
-						"percent": "21.0%"
-					}
-				],
-				"total": "200.00"
-			},
-			{
-				"i": 2,
-				"quantity": "50",
-				"item": {
-					"name": "Branded Mugs",
-					"price": "7.50",
-					"meta": {
-						"product": "goods"
-					}
-				},
-				"sum": "375.00",
-				"taxes": [
-					{
-						"cat": "VAT",
-						"rate": "standard",
-						"percent": "21.0%"
-					}
-				],
-				"total": "375.00"
-			}
-		],
-		"totals": {
-			"sum": "575.00",
-			"tax_included": "99.79",
-			"total": "475.21",
-			"taxes": {
-				"categories": [
-					{
-						"code": "VAT",
-						"rates": [
-							{
-								"key": "standard",
-								"base": "475.21",
-								"percent": "21.0%",
-								"amount": "99.79"
-							}
-						],
-						"base": "475.21",
-						"amount": "99.79"
-					}
-				],
-				"sum": "99.79"
-			},
-			"tax": "0.00",
-			"total_with_tax": "475.21",
-			"payable": "475.21"
-		},
-		"notes": [
-			{
-				"key": "legal",
-				"src": "reverse-charge",
-				"text": "Reverse Charge / Inversión del sujeto pasivo."
-			}
-		]
-	},
-	"sigs": [
-		"eyJhbGciOiJFUzI1NiIsImtpZCI6IjgzNDdjNGQwLTdjZDQtNGMyMi05OWY2LTg5MWI1NDRlMWMxZCJ9.eyJ1dWlkIjoiMjUyMTQyMzUtZWJjNy0xMWVjLTkxNTUtZTJjMDYyNDcwM2Q5IiwiZGlnIjp7ImFsZyI6InNoYTI1NiIsInZhbCI6IjJjMTZhMDA5YzU3NWNlYjg2MmU5NGE2ZGE5YThiOTAxNDBiNmFhYjZlYTkzYjM3ZDE1NGZlMTk2NjE4ZTA5Y2QifX0.HnyALAZjZgX5kyUcxwIImw0piBEooDmmitHbaWpB7OKzlkavJz0dzb04GZMxISv6ySpZIj_fcTnTut3Zw0vtPw"
-	]
-}
