--- samples/output/message.json
@@ -1,18 +1,0 @@
-{
-	"$schema": "https://gobl.org/draft-0/envelope",
-	"head": {
-		"uuid": "6869c8be-ebc8-11ec-93e4-e2c0624703d9",
-		"dig": {
-			"alg": "sha256",
-			"val": "45ac3115c8569a1789e58af8d0dc91ef3baa1fb71daaf38f5aef94f82b4d0033"
-		}
-	},
-	"doc": {
-		"$schema": "https://gobl.org/draft-0/note/message",
-		"title": "Test Message",
-		"content": "We hope you like this test message!"
-	},
-	"sigs": [
-		"eyJhbGciOiJFUzI1NiIsImtpZCI6ImEzMjU5MTNiLTg4NTAtNGVmNy1hMjE5LWNkYmQ5MmMwMjk0ZCJ9.eyJ1dWlkIjoiNjg2OWM4YmUtZWJjOC0xMWVjLTkzZTQtZTJjMDYyNDcwM2Q5IiwiZGlnIjp7ImFsZyI6InNoYTI1NiIsInZhbCI6IjQ1YWMzMTE1Yzg1NjlhMTc4OWU1OGFmOGQwZGM5MWVmM2JhYTFmYjcxZGFhZjM4ZjVhZWY5NGY4MmI0ZDAwMzMifX0.sPhIviD4GrZYqNKC7N77PGriiNSzPi6g6vf5CanDYseGkkemJdnVPR-X6t-_ZHn50aYRv6zYsHMu8QM2Sxtn7g"
-	]
-}
